Added requirements for installer and renamed module (#3)
