+++ README.md
@@ -8,7 +8,7 @@ Getting started is fairly straight forward.
 
 1. Clone the repo to the directory you want it to run in. 
 2. Install the requirements `python -m pip install -r requirements.txt`
-3. Run the OpenSearch docker compose
+3. Run the OpenSearch docker compose (if not using an already stood up instance)
     - `cd infra`
     - `docker-compose --env-file env up`
 4. Create a .env file in the root of the repo with the required variables
@@ -17,5 +17,3 @@ Getting started is fairly straight forward.
     - `crontab -e 0 * * * *  cd <working dir> && <path to python> <path to main.py>`
 6. *Optional:* Setup translation
     - Setup the authentication for Googles Translation API: https://cloud.google.com/translate/docs/setup
-
-opensearchDashboards.branding.useExpandedHeader: false
