+++ community_pulse/main.py
@@ -1,23 +1,27 @@
 from os import getenv
 import dotenv
-from opensearchpy import OpenSearch
-from cluster_init import enforce_index_templates, create_meta_indices
-from util import parse_config, initialize_opensearch_client, init_logging
-import twitter
+from community_pulse.cluster_init import enforce_index_templates, create_meta_indices
+from community_pulse.util import parse_config, initialize_opensearch_client, init_logging
+import community_pulse.twitter
 import logging
 from logging.handlers import RotatingFileHandler
 import time
+import argparse
 
 logger = logging.getLogger('community-pulse')
 dotenv.load_dotenv()
 
-if __name__=="__main__":
+def main():
+    parser = argparse.ArgumentParser(description='Community-Pulse is a project for collecting and aggregating community data.')
+    args = parser.parse_args()
+
+
     full_config = parse_config('./test_config/community_pulse.yml')
     init_logging(logger, full_config['settings']['log_level'])
     initialize_opensearch_client(full_config['settings']['opensearch'])
 
     jobs_map = {
-        'twitter': twitter.gen_twitter_executor()
+        'twitter': community_pulse.twitter.gen_twitter_executor()
     }
 
     # Should make this a little smarter
@@ -31,3 +35,7 @@ if __name__=="__main__":
         curr_job(**config)
         end = time.time()
         logger.debug(f"Finshed {job} in: {end - start}s")
+
+
+if __name__=="__main__":
+    main()
