+++ pyproject.toml
@@ -1,0 +1,26 @@
+[build-system]
+requires = ["hatchling", "tweepy", "opensearch-py", "google-cloud-translate", "iso-639", "pyyaml", "jsonpath-ng"]
+build-backend = "hatchling.build"
+
+[project]
+name = "community-pulse"
+version = "0.0.1"
+authors = [
+  { name="David Tippett" },
+]
+description = "Community Pulse is a collection of tools for community monitoring"
+readme = "README.md"
+license = { file="LICENSE" }
+requires-python = ">=3.7"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: Apache Software License",
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/dtaivpp/community-pulse"
+"Bug Tracker" = "https://github.com/dtaivpp/community-pulse/issues"
+
+[project.scripts]
+pulse = "community_pulse.main:main"
